requirements.txt
streamlit
pypdf2
langchain
faiss-cpu
langchain-community == 0.3.29
tiktoken == 0.11.0
python-dotenv